requirements.txt
pandas==2.2.2
numpy==2.0.0
sqlalchemy<2.0
psycopg2-binary==2.9.9
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2